+++ lib/library/widgets/hoverable_area.dart
@@ -8,15 +8,18 @@ typedef HoverableWidgetBuilder = Widget Function(
 class HoverableArea extends HookWidget {
   const HoverableArea({
     required this.builder,
+    this.clickable = true,
     final Key? key,
   }) : super(key: key);
   final HoverableWidgetBuilder builder;
+  final bool clickable;
+
   @override
   Widget build(final BuildContext context) {
     final hovered = useIsBool();
 
     return MouseRegion(
-      cursor: SystemMouseCursors.click,
+      cursor: clickable ? SystemMouseCursors.click : SystemMouseCursors.basic,
       onEnter: (final _) => hovered.value = true,
       onExit: (final _) => hovered.value = false,
       child: builder(context, hovered.value),