+++ lib/library/widgets/special_emoji_grid.dart
@@ -4,15 +4,21 @@ class SpecialEmojiPopup extends HookWidget {
   const SpecialEmojiPopup({
     required final this.controller,
     required final this.focusNode,
+    required final this.onShowEmojiKeyboard,
     final Key? key,
   }) : super(key: key);
   final TextEditingController controller;
   final FocusNode focusNode;
+  final VoidCallback onShowEmojiKeyboard;
 
   @override
   Widget build(final BuildContext context) {
-    if (!isNativeDesktop) return const SizedBox();
-
+    if (!isNativeDesktop && !kIsWeb) {
+      return IconButton(
+        onPressed: onShowEmojiKeyboard,
+        icon: const Icon(Icons.emoji_flags_rounded),
+      );
+    }
     final emojiInserter = EmojiInserter.use(
       controller: controller,
       focusNode: focusNode,
@@ -40,10 +46,16 @@ class SpecialEmojisGrid extends StatelessWidget {
   final bool hideBorder;
   @override
   Widget build(final BuildContext context) {
+    final emojiStyle = isNativeDesktop && Platform.isMacOS
+        ? null
+        : Theme.of(context).textTheme.bodyText2?.copyWith(
+              fontFamily: 'NotoColorEmoji',
+            );
     Widget buildEmojiButton(final Emoji emoji) {
       return EmojiButton(
         key: ValueKey(emoji),
         emoji: emoji,
+        style: emojiStyle,
         onPressed: () => onChanged(emoji),
       );
     }