+++ lib/pack_settings/features_widgets/characters_limit.dart
@@ -24,9 +24,11 @@ class CharactersLimitSetting extends HookWidget {
   @override
   Widget build(final BuildContext context) {
     final theme = Theme.of(context);
+    final dark = theme.brightness == Brightness.dark;
     final settings = SettingsStateScope.of(context);
     final initialText = getInitialLimit(settings: settings);
     final controller = useTextEditingController(text: initialText);
+
     final state = useCharactersLimitSettingStateState(
       context: context,
       note: note,
@@ -34,44 +36,170 @@ class CharactersLimitSetting extends HookWidget {
       controller: controller,
     );
 
-    return Row(
+    Widget otherButton;
+    // TODO(arenukvern): refactor to separate widget
+
+    if (state.noLimitIsSet) {
+      otherButton = Padding(
+        padding: const EdgeInsets.only(right: 8.0),
+        child: HoverableButton(
+          onPressed: () => state.setLimit(
+            0,
+            updateController: true,
+            zeroEqualNull: false,
+          ),
+          child: AnimatedBuilder(
+            animation: settings,
+            builder: (final context, final _) =>
+                Text(S.current.charactersUnlimited),
+          ),
+        ),
+      );
+    } else {
+      otherButton = Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const SizedBox(width: 12),
+          SizedBox(
+            width: 60,
+            child: TextField(
+              controller: controller,
+              keyboardType: TextInputType.number,
+              autocorrect: false,
+              keyboardAppearance: theme.brightness,
+              enableSuggestions: false,
+              inputFormatters: [FilteringTextInputFormatter.digitsOnly],
+              decoration: const InputDecoration()
+                  .applyDefaults(theme.inputDecorationTheme)
+                  .copyWith(
+                    hintText: S.current.charactersUnlimited,
+                  ),
+            ),
+          ),
+          const SizedBox(width: 2),
+          HoverableButton(
+            onPressed: controller.text.isEmpty ? null : state.onClearLimit,
+            child: const Icon(
+              CupertinoIcons.clear,
+              size: 14,
+            ),
+          ),
+        ],
+      );
+    }
+    // TODO(arenukvern): refactor to separate widget
+
+    SvgGenImage vkIcon;
+    if (state.isVkLimit) {
+      vkIcon = Assets.icons.vkLogoBlue;
+    } else {
+      vkIcon = dark ? Assets.icons.vkLogoWhite : Assets.icons.vkLogoBlack;
+    }
+    // TODO(arenukvern): refactor to separate widget
+
+    AssetGenImage instagramIcon;
+    if (state.isInstagramLimit) {
+      instagramIcon = Assets.icons.instagramLogoColorful;
+    } else {
+      instagramIcon = Assets.icons.instagramLogoBlack;
+    }
+    // TODO(arenukvern): refactor to separate widget
+
+    SvgGenImage twitterIcon;
+    if (state.isTwitterLimit) {
+      twitterIcon = Assets.icons.twitterLogoBlue;
+    } else {
+      twitterIcon =
+          dark ? Assets.icons.twitterLogoWhite : Assets.icons.twitterLogoBlack;
+    }
+    // TODO(arenukvern): refactor to separate widget
+
+    AssetGenImage facebookIcon;
+    if (state.isFacebookLimit) {
+      facebookIcon = Assets.icons.fbLogoBlue;
+    } else {
+      facebookIcon = dark ? Assets.icons.fbLogoWhite : Assets.icons.fbLogoBlack;
+    }
+    // TODO(arenukvern): refactor to separate widget
+    SvgGenImage discordIcon;
+    if (state.isDiscordLimit) {
+      discordIcon = Assets.icons.discordLogoBlue;
+    } else {
+      discordIcon =
+          dark ? Assets.icons.discordLogoWhite : Assets.icons.discordLogoBlack;
+    }
+
+    return Wrap(
+      spacing: 14,
+      runSpacing: 14,
       children: [
-        AppIconButton(
-          onPressed: state.onSetInstagramLimit,
-          child: Padding(
-            padding: const EdgeInsets.symmetric(
-              vertical: 3,
-              horizontal: 3,
-            ),
-            child: Assets.icons.instagramLogo.image(
-              width: 18,
-              height: 18,
-              color: theme.textTheme.bodyText2?.color,
-            ),
+        CharactersLimitButton(
+          onTap: state.onSetInstagramLimit,
+          child: instagramIcon.image(
+            width: 18,
+            height: 18,
+            color: state.isInstagramLimit
+                ? null
+                : theme.textTheme.bodyText2?.color,
           ),
         ),
-        const SizedBox(width: 10),
-        Expanded(
-          child: TextField(
-            controller: controller,
-            keyboardType: TextInputType.number,
-            autocorrect: false,
-            keyboardAppearance: theme.brightness,
-            enableSuggestions: false,
-            inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-            decoration: const InputDecoration()
-                .applyDefaults(theme.inputDecorationTheme)
-                .copyWith(
-                  hintText: S.current.charactersUnlimited,
-                ),
+        CharactersLimitButton(
+          onTap: state.onSetTwitterLimit,
+          child: twitterIcon.svg(
+            width: 16,
+            height: 16,
+            color: state.isTwitterLimit
+                ? AppColors.twitterBlue
+                : theme.textTheme.bodyText2?.color,
           ),
         ),
-        const SizedBox(width: 2),
-        AppIconButton(
-          onPressed: controller.text.isEmpty ? null : state.onClearLimit,
-          child: const Icon(CupertinoIcons.clear),
+        CharactersLimitButton(
+          onTap: state.onSetFacebookLimit,
+          child: facebookIcon.image(
+            width: 16,
+            height: 16,
+          ),
         ),
+        CharactersLimitButton(
+          onTap: state.onSetDiscordLimit,
+          child: discordIcon.svg(
+            width: 16,
+            height: 16,
+          ),
+        ),
+        CharactersLimitButton(
+          onTap: state.onSetVkLimit,
+          child: vkIcon.svg(
+            width: 16,
+            height: 16,
+          ),
+        ),
+        otherButton,
       ],
+    );
+  }
+}
+
+class CharactersLimitButton extends StatelessWidget {
+  const CharactersLimitButton({
+    required this.onTap,
+    required this.child,
+    final Key? key,
+  }) : super(key: key);
+  final VoidCallback onTap;
+  final Widget child;
+
+  @override
+  Widget build(final BuildContext context) {
+    return HoverableButton(
+      onPressed: onTap,
+      child: Padding(
+        padding: const EdgeInsets.symmetric(
+          vertical: 3,
+          horizontal: 3,
+        ),
+        child: child,
+      ),
     );
   }
 }