+++ lib/pack_settings/pack_settings.dart
@@ -1,6 +1,7 @@
 library pack_settings;
 
 import 'dart:async';
+import 'dart:math' as math;
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -11,6 +12,7 @@ import 'package:lastanswer/abstract/abstract.dart';
 import 'package:lastanswer/generated/assets.gen.dart';
 import 'package:lastanswer/generated/l10n.dart';
 import 'package:lastanswer/library/hooks/hooks.dart';
+import 'package:lastanswer/library/theme/theme.dart';
 import 'package:lastanswer/library/widgets/widgets.dart';
 import 'package:lastanswer/pack_app/pack_app.dart';
 import 'package:lastanswer/pack_note/pack_note.dart';